+++ app/src/main/java/woowacourse/shopping/ShoppingApplication.kt
@@ -1,13 +1,12 @@
 package woowacourse.shopping
 
 import android.app.Application
+import woowacourse.shopping.annotation.Qualifier
 import woowacourse.shopping.data.CartProductDao
 import woowacourse.shopping.data.DatabaseCartRepository
 import woowacourse.shopping.data.DefaultProductRepository
 import woowacourse.shopping.data.InMemoryCartRepository
 import woowacourse.shopping.data.ShoppingDatabase
-import woowacourse.shopping.di.AppContainer
-import woowacourse.shopping.di.annotation.Qualifier
 
 class ShoppingApplication : Application() {
     lateinit var appContainer: AppContainer
@@ -20,8 +19,17 @@ class ShoppingApplication : Application() {
             CartProductDao::class,
             ShoppingDatabase.getInstance(this)::cartProductDao,
         )
-        appContainer.addQualifier(Qualifier("defaultProductRepository"), DefaultProductRepository::class)
-        appContainer.addQualifier(Qualifier("databaseCartRepository"), DatabaseCartRepository::class)
-        appContainer.addQualifier(Qualifier("inMemoryCartRepository"), InMemoryCartRepository::class)
+        appContainer.addQualifier(
+            Qualifier("defaultProductRepository"),
+            DefaultProductRepository::class,
+        )
+        appContainer.addQualifier(
+            Qualifier("databaseCartRepository"),
+            DatabaseCartRepository::class,
+        )
+        appContainer.addQualifier(
+            Qualifier("inMemoryCartRepository"),
+            InMemoryCartRepository::class,
+        )
     }
 }
