+++ app/src/main/java/woowacourse/shopping/data/DatabaseCartRepository.kt
@@ -1,8 +1,8 @@
 package woowacourse.shopping.data
 
+import woowacourse.shopping.annotation.SingleInstance
 import woowacourse.shopping.data.mapper.toDomain
 import woowacourse.shopping.data.mapper.toEntity
-import woowacourse.shopping.di.annotation.SingleInstance
 import woowacourse.shopping.model.CartProduct
 import woowacourse.shopping.model.Product
 import woowacourse.shopping.repository.CartRepository
