+++ app/src/main/java/woowacourse/shopping/data/InMemoryCartRepository.kt
@@ -1,6 +1,6 @@
 package woowacourse.shopping.data
 
-import woowacourse.shopping.di.annotation.SingleInstance
+import woowacourse.shopping.annotation.SingleInstance
 import woowacourse.shopping.model.CartProduct
 import woowacourse.shopping.model.Product
 import woowacourse.shopping.repository.CartRepository
