+++ di/build.gradle.kts
@@ -1,0 +1,13 @@
+plugins {
+    id("java-library")
+    id("org.jetbrains.kotlin.jvm")
+}
+
+java {
+    sourceCompatibility = JavaVersion.VERSION_1_7
+    targetCompatibility = JavaVersion.VERSION_1_7
+}
+
+dependencies {
+    implementation("org.jetbrains.kotlin", "kotlin-reflect", "1.8.21")
+}
